renamed dc2 to data-collector. #763
