+++ buildout.cfg
@@ -57,7 +57,6 @@ dc_cmd = ${buildout:directory}/bin/data-collector
 sm_cmd = ${buildout:directory}/bin/service-monitor
 jd_cmd = ${buildout:directory}/bin/job-dispatcher
 ztask_cmd = ${buildout:directory}/bin/django ztaskd --noreload -l DEBUG -f ${supervisord-conf:logdir}/ztask.log
-dc2_cmd = ${buildout:directory}/bin/dc2
 
 [django-settings-conf]
 rootdir = ${buildout:directory}/src/rockstor
