+++ prod-buildout.cfg
@@ -55,7 +55,6 @@ replicad_cmd = ${buildout:depdir}/bin/replicad
 dc_cmd = ${buildout:depdir}/bin/data-collector
 sm_cmd = ${buildout:depdir}/bin/service-monitor
 ztask_cmd = ${buildout:depdir}/bin/django ztaskd --noreload --replayfailed -f ${supervisord-conf:logdir}/ztask.log
-dc2_cmd = ${buildout:depdir}/bin/dc2
 input = ${buildout:directory}/conf/supervisord-prod.conf.in
 
 [django-settings-conf]
